DynamoDB Manager usage rebuilt
Object "Topic" is a representaion of a row from DynamoDB. Each instance attribte is a row attribute. If instance of Topic has been updated and method of db manager "save" was called, the row in Dynamo will be updated. Also some methods for do CRUD operations was added to DynamoDB class